+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 - [#49] Microsite for documentation ([https://doc.norcane.com/headroom/latest/](https://doc.norcane.com/headroom/latest/))
 - [#51] Allow template variable values to be templates itself
 - [#52] Postprocessing functions for _license headers_
-- Bump _LTS Haskell_ to `16.0`
+- Various bugfixes and performance improvements
+- Bump _LTS Haskell_ to `16.1`
 
 ## 0.2.2.1 (released 2020-05-08)
 - Minor improves in generated `.headroom.yaml` in `init` mode